Use BUILD_SHARED_LIBS, not JPEGXL_STATIC
JPEGXL_STATIC means "Build tools as static binaries" and is not intended
for library builds. In fact it causes strange build issues on FreeBSD
by trying to link libc (!) statically.
It does not disable the building of shared libraries though.

In 0.6.1, there is no way to disable the building of shared libraries.
However in the main branch BUILD_SHARED_LIBS is already implemented:
https://github.com/libjxl/libjxl/pull/575